+++ CHANGELOG.md
@@ -1,10 +1,16 @@
-# 1.7.2
+# 1.8.0
+
+### Added
+
+- Gave Floo Bricks an item
+- Added an EMI recipe for how to make Floo Bricks
 
 ### Changed
 
 - Ender Letters now uses the username cache instead of the list of online players
 	- Now you can send Letters to players that are offline, as long as they've joined the server before
 - Added more Patchouli pages
+- Floo Fireplaces no longer cancel your message when you teleport to another Fireplace
 
 ### Removed